# swagger/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "My Address Book API",
    "version": "1.0.0",
    "description": "<p>API <strong>REStfull</strong> pour gerer un carnet d'adresses</p>"
  },
  "servers": [
    {
      "url": "https://api.myaddressesbook.com",
      "description": "Serveur de production"
    },
    {
      "url": "http://localhost:3000",
      "description": "Serveur local"
    }
  ],
  "tags": [
    {
      "name": "Utilisateurs",
      "description": "Tout ce qui conerner les utilisateurs"
    },
    {
      "name": "Contacts",
      "description": "Tout ce qui conerner les contact"
    }
  ],
  "paths": {
    "/users": {
      "get": {
        "tags": ["Utilisateurs"],
        "summary": "Recupere la liste des Utilisateurs",
        "description": "Voire la liste des Utilisateurs de l'API",
        "operationId": "getUsers",
        "responses": {
          "200": {
            "description": "Voire la liste des Utilisateurs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArrayOfUtilisateurs"
                }
              }
            }
          }
        }
      }
    },
    "/users/register": {
      "post": {
        "summary": "Enregistre un nouvel utilisateur",
        "description": "Enregistre un nouvel utilisateur avec les informations fournies",
        "tags": ["Utilisateurs"],
        "operationId": "registerUser",
        "requestBody": {
          "description": "Les informations du nouvel utilisateur",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Utilisateurs"
              }
            }
          }
        }
      }
    },
    "/users/login": {
      "post": {
        "tags": ["Utilisateurs"],
        "summary": "Connexion d'un utilisateur",
        "description": "Effectue la connexion d'un utilisateur avec les informations fournies",
        "operationId": "loginUser",
        "requestBody": {
          "description": "Les informations de connexion",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Utilisateurs"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Connexion reussie",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Utilisateurs"
                }
              }
            }
          }
        }
      }
    },
    "/users/logout": {
      "delete": {
        "tags": ["Utilisateurs"],
        "summary": "Deconnexion d'un utilisateur",
        "description": "Effectue la deconnexion d'un utilisateur",
        "operationId": "logoutUser",
        "parameters": [
          {
            "name": "accessToken",
            "in": "cookie",
            "description": "Token d'accès pour authentification",
            "required": true,
            "schema": {
              "type": "string",
              "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deconnexion reussie"
          },
          "401": {
            "description": "Deconnexion echouee"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": ["Utilisateurs"],
        "summary": "Obtenir un utilisateur",
        "description": "Recupere un utilisateur selon son ID",
        "operationId": "getUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identifiant de l'utilisateur",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Renvoie un utilisateur",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Utilisateurs"
                }
              }
            }
          },
          "404": {
            "desription": "Renvoie une erreur, aucun element trouve"
          }
        }
      },
      "patch": {
        "tags": ["Utilisateurs"],
        "summary": "Mettre a jour un utilisateur",
        "description": "Mettre a jour un utilisateur selon son ID",
        "operationId": "updateUser",
        "requestBody": {
          "required": true,
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "description": "ID de l'utilisateur à mettre à jour",
              "schema": {
                "type": "string"
              }
            }
          ],
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        }
      }
    },
    "/contacts": {
      "get": {
        "summary": "Récupérer tous les contacts d'un utilisateur connecté",
        "tags": ["Contacts"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "description": "Token d'authentification au format Bearer",
            "schema": {
              "type": "string",
              "example": "Bearer your_token_here"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des contacts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "example": "61a5b3f58e8b9f001c8d4567"
                      },
                      "name": { "type": "string", "example": "Jane Doe" },
                      "email": {
                        "type": "string",
                        "example": "jane.doe@example.com"
                      },
                      "phone": { "type": "string", "example": "555-1234" }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Non autorisé" }
        }
      },
      "post": {
        "tags": ["Contacts"],
        "summary": "Creer un contact",
        "description": "Creer un contact en donnant toutes les informations necessaires",
        "operationId": "createContact",
        "requestBody": {
          "description": "contenu complet d'un contact",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Contacts"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Articcontactle creer avec success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contacts"
                }
              }
            }
          }
        }
      }
    }
  },

  "components": {
    "schemas": {
      "Utilisateurs": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "example": {
          "name": "blabla",
          "email": "blabla@gmail.com",
          "password": "123456"
        }
      },
      "ArrayOfUtilisateurs": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/Utilisateurs"
        },
        "example": [
          {
            "id": 1,
            "name": "Didi",
            "email": "didi@gmail.com"
          },
          {
            "id": 2,
            "username": "Bidi",
            "email": "bibi@gmail.com"
          }
        ]
      },
      "Contacts": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "iri-reference",
            "default": "/users/{id}"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          }
        },
        "example": {
          "userId": "/users/1",
          "name": "nolan",
          "email": "nolan@gmail.com",
          "phone": "5066551887"
        }
      },
      "ArrayOfContacts": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/Contacts"
        },
        "example": [
          {
            "userId": 1,
            "name": "nolan",
            "email": "nolan@gmail.com",
            "phone": "5066551887"
          },
          {
            "userId": 2,
            "username": "alan",
            "email": "alan@gmail.com",
            "phone": "5056051907"
          }
        ]
      }
    }
  }
}